pyproject.toml
[tool.poetry]
name = "scanli"
version = "0.1.1"
description = "CLI tool which scans project dependencies for OSS licenses."
authors = ["Chris Lawlor <lawlor.chris@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/chrislawlor/scanli"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Utilities"

]

[tool.poetry.scripts]
scanli = "scanli:main"

[tool.poetry.dependencies]
python = "^3.10"
requirements-parser = "^0.5.0"
rich = "^12.6.0"
cache-to-disk = "^2.0.0"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
types-requests = "^2.28.11.2"
black = "^22.10.0"
flake8 = "^5.0.4"
mypy = "^0.982"
ipython = "^8.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
